- gringo compiles on 64bit architectures now   (at least on jersey)
git-svn-id: https://potassco.svn.sourceforge.net/svnroot/potassco/branches/gringo2@409 3e02f596-3131-0410-8352-e04eeaec37cc

+++ lib/gringo/src/funcsymbol.cpp
@@ -49,7 +49,7 @@ size_t FuncSymbol::getHash() const
 	Value::VectorHash t;
 	hash = t(args_);
 
-	hash = (hash << 4) + int(name_);
+	hash = (hash << 4) + size_t(name_);
 	if((x = hash & 0xF0000000L) != 0)
 	{
 		hash ^= (x >> 24);
